# change AndExpr to contain arbitrary many predicates