--- a/go/vt/sqlparser/ast_clone.go
+++ b/go/vt/sqlparser/ast_clone.go
@@ -744,8 +744,7 @@ func CloneRefOfAndExpr(n *AndExpr) *AndExpr {
 		return nil
 	}
 	out := *n
-	out.Left = CloneExpr(n.Left)
-	out.Right = CloneExpr(n.Right)
+	out.Predicates = CloneSliceOfExpr(n.Predicates)
 	return &out
 }
 
@@ -4369,6 +4368,18 @@ func CloneSliceOfIdentifierCI(n []IdentifierCI) []IdentifierCI {
 	return res
 }
 
+// CloneSliceOfExpr creates a deep clone of the input.
+func CloneSliceOfExpr(n []Expr) []Expr {
+	if n == nil {
+		return nil
+	}
+	res := make([]Expr, len(n))
+	for i, x := range n {
+		res[i] = CloneExpr(x)
+	}
+	return res
+}
+
 // CloneSliceOfTxAccessMode creates a deep clone of the input.
 func CloneSliceOfTxAccessMode(n []TxAccessMode) []TxAccessMode {
 	if n == nil {
@@ -4454,18 +4465,6 @@ func CloneSliceOfRefOfVariable(n []*Variable) []*Variable {
 	res := make([]*Variable, len(n))
 	for i, x := range n {
 		res[i] = CloneRefOfVariable(x)
-	}
-	return res
-}
-
-// CloneSliceOfExpr creates a deep clone of the input.
-func CloneSliceOfExpr(n []Expr) []Expr {
-	if n == nil {
-		return nil
-	}
-	res := make([]Expr, len(n))
-	for i, x := range n {
-		res[i] = CloneExpr(x)
 	}
 	return res
 }