--- a/go/vt/sqlparser/ast_visit.go
+++ b/go/vt/sqlparser/ast_visit.go
@@ -811,10 +811,8 @@ func VisitRefOfAndExpr(in *AndExpr, f Visit) error {
 	if cont, err := f(in); err != nil || !cont {
 		return err
 	}
-	for _, el := range in.Predicates {
-		if err := VisitExpr(el, f); err != nil {
-			return err
-		}
+	if err := VisitExprs(in.Predicates, f); err != nil {
+		return err
 	}
 	return nil
 }