… go/vt/sqlparser/analyzer_test.go
… go/vt/sqlparser/precedence_test.go
… go/vt/sqlparser/predicate_rewriting_test.go
… go/vt/sqlparser/testdata/select_cases.txt
… go/vt/sqlparser/utils.go
… go/vt/vtctl/workflow/materializer.go
… go/vt/vtctl/workflow/stream_migrator.go
… go/vt/vtctl/workflow/traffic_switcher.go
… go/vt/vtctl/workflow/utils.go
… go/vt/vtctl/workflow/vexec/query_plan.go
… go/vt/vtctl/workflow/vexec/query_planner.go
… go/vt/vtctl/workflow/vexec/query_planner_test.go
… go/vt/vtgate/evalengine/translate.go
… go/vt/vtgate/planbuilder/operators/ast_to_op.go
… go/vt/vtgate/planbuilder/operators/phases.go
… go/vt/vtgate/planbuilder/operators/querygraph.go
… go/vt/vtgate/planbuilder/operators/subquery.go
… go/vt/vtgate/planbuilder/operators/update.go
… go/vt/vtgate/planbuilder/predicate_rewrite_test.go
… go/vt/vtgate/planbuilder/testdata/filter_cases.json
… go/vt/vtgate/planbuilder/testdata/foreignkey_cases.json
… go/vt/vtgate/planbuilder/testdata/foreignkey_checks_on_cases.json
… go/vt/vtgate/planbuilder/testdata/select_cases.json
… go/vt/vtgate/semantics/early_rewriter.go
… go/vt/vtgate/semantics/early_rewriter_test.go
… go/vt/vtgate/semantics/scoper.go
… go/vt/vtgate/semantics/semantic_table.go
… go/vt/vtgate/simplifier/expression_simplifier.go
… go/vt/vttablet/tabletserver/vstreamer/planbuilder.go
… go/vt/wrangler/materializer.go