## Description
During planning, we often want to have predicates as a slice and not in a tree structure. This PR changes the `AndExpr` to contain a slice if inner predicates instead of `Left/Right` fields.

This PR also moves the checking for constant values in predicates to the planning stage. No need to do these rewrites if the whole query will be sent to MySQL.

## Summary of Performance Changes:

The recent changes introduced in this PR have resulted in mixed performance outcomes across various benchmarks:

1.	Plan Time (sec/op):
 * There is a slight overall increase in execution time with a geometric mean change of +0.48%.
 * The OLTP/Gen4 series saw notable increases in execution time, ranging from +18.33% to +23.20%.
 * Conversely, the TPCC/Gen4 series experienced performance improvements, with execution times decreasing by up to -6.52%.
 * The Planner benchmarks mostly saw increases in execution time, with the largest increase of +8.41% in Planner/from_cases.json-gen4-20.
2.	Memory Usage (B/op):
 * Overall memory usage decreased slightly, with a geometric mean reduction of -2.18%.
 * Significant reductions were observed in Planner/filter_cases.json-gen4 and Planner/filter_cases.json-gen4left2right, both decreasing by over -20%.
 * However, memory usage slightly increased in the OLTP/Gen4 series by approximately +3.6%.
3.	Memory Allocations (allocs/op):
 * The total number of memory allocations showed a minor decrease of -0.65%.
 * Notable increases in allocations occurred in the SelectVsDML/DML_(random_sample,_N=32) benchmark (+11.34%).
 * However, the SelectVsDML/Select_(random_sample,_N=32) benchmark saw a significant reduction in allocations by -17.33%.

These results reflect targeted optimizations and regressions across different query plans and scenarios, highlighting areas where the changes have improved or slightly degraded performance.

Details:
```
                                            │  ../before  │                after                 │
                                            │   sec/op    │    sec/op     vs base                │
OLTP/Gen4-20                                  221.3µ ± 1%    261.9µ ± 1%  +18.33% (p=0.000 n=10)
OLTP/Gen4Greedy-20                            220.9µ ± 0%    270.1µ ± 3%  +22.26% (p=0.000 n=10)
OLTP/Gen4Left2Right-20                        220.1µ ± 2%    271.2µ ± 1%  +23.20% (p=0.000 n=10)
TPCC/Gen4-20                                  2.259m ± 1%    2.189m ± 1%   -3.12% (p=0.000 n=10)
TPCC/Gen4Greedy-20                            2.255m ± 0%    2.170m ± 2%   -3.77% (p=0.000 n=10)
TPCC/Gen4Left2Right-20                        2.225m ± 0%    2.080m ± 1%   -6.52% (p=0.000 n=10)
TPCH/Gen4-20                                  11.16m ± 1%    11.18m ± 1%        ~ (p=0.143 n=10)
TPCH/Gen4Greedy-20                            11.17m ± 0%    11.27m ± 1%        ~ (p=0.353 n=10)
TPCH/Gen4Left2Right-20                        9.988m ± 0%   10.022m ± 1%        ~ (p=0.739 n=10)
Planner/from_cases.json-gen4-20               7.336m ± 1%    7.953m ± 2%   +8.41% (p=0.000 n=10)
Planner/from_cases.json-gen4left2right-20     6.876m ± 1%    7.376m ± 2%   +7.26% (p=0.000 n=10)
Planner/filter_cases.json-gen4-20             181.8m ± 0%    155.2m ± 1%  -14.61% (p=0.000 n=10)
Planner/filter_cases.json-gen4left2right-20   181.9m ± 0%    155.0m ± 1%  -14.82% (p=0.000 n=10)
Planner/large_cases.json-gen4-20              484.6µ ± 1%    500.1µ ± 1%   +3.20% (p=0.000 n=10)
Planner/large_cases.json-gen4left2right-20    301.2µ ± 1%    306.7µ ± 1%   +1.80% (p=0.002 n=10)
Planner/aggr_cases.json-gen4-20               13.04m ± 0%    14.09m ± 1%   +8.13% (p=0.000 n=10)
Planner/aggr_cases.json-gen4left2right-20     12.46m ± 1%    13.42m ± 1%   +7.68% (p=0.000 n=10)
Planner/select_cases.json-gen4-20             10.39m ± 1%    11.17m ± 1%   +7.52% (p=0.000 n=10)
Planner/select_cases.json-gen4left2right-20   10.13m ± 2%    10.91m ± 1%   +7.66% (p=0.000 n=10)
Planner/union_cases.json-gen4-20              3.732m ± 6%    4.009m ± 1%   +7.42% (p=0.000 n=10)
Planner/union_cases.json-gen4left2right-20    3.710m ± 2%    3.990m ± 1%   +7.55% (p=0.000 n=10)
SemAnalysis-20                                62.98m ± 3%    29.67m ± 1%  -52.89% (p=0.000 n=10)
SelectVsDML/DML_(random_sample,_N=32)-20      946.7µ ± 2%   1093.3µ ± 1%  +15.49% (p=0.000 n=10)
SelectVsDML/Select_(random_sample,_N=32)-20   1.948m ± 5%    1.843m ± 1%   -5.40% (p=0.000 n=10)
geomean                                       4.220m         4.240m        +0.48%

                                            │   ../before   │                after                 │
                                            │     B/op      │     B/op      vs base                │
OLTP/Gen4-20                                   144.9Ki ± 0%   150.1Ki ± 0%   +3.59% (p=0.000 n=10)
OLTP/Gen4Greedy-20                             144.9Ki ± 0%   150.1Ki ± 0%   +3.58% (p=0.000 n=10)
OLTP/Gen4Left2Right-20                         144.9Ki ± 0%   150.1Ki ± 0%   +3.58% (p=0.000 n=10)
TPCC/Gen4-20                                   1.056Mi ± 0%   1.066Mi ± 0%   +0.96% (p=0.000 n=10)
TPCC/Gen4Greedy-20                             1.056Mi ± 0%   1.066Mi ± 0%   +0.96% (p=0.000 n=10)
TPCC/Gen4Left2Right-20                         1.045Mi ± 0%   1.055Mi ± 0%   +0.95% (p=0.000 n=10)
TPCH/Gen4-20                                   5.328Mi ± 0%   5.290Mi ± 0%   -0.70% (p=0.000 n=10)
TPCH/Gen4Greedy-20                             5.328Mi ± 0%   5.289Mi ± 0%   -0.72% (p=0.000 n=10)
TPCH/Gen4Left2Right-20                         4.806Mi ± 0%   4.775Mi ± 0%   -0.64% (p=0.000 n=10)
Planner/from_cases.json-gen4-20                4.503Mi ± 0%   4.521Mi ± 0%   +0.40% (p=0.000 n=10)
Planner/from_cases.json-gen4left2right-20      4.314Mi ± 0%   4.332Mi ± 0%   +0.43% (p=0.000 n=10)
Planner/filter_cases.json-gen4-20              22.39Mi ± 0%   17.75Mi ± 0%  -20.72% (p=0.000 n=10)
Planner/filter_cases.json-gen4left2right-20    22.31Mi ± 0%   17.68Mi ± 0%  -20.76% (p=0.000 n=10)
Planner/large_cases.json-gen4-20               268.9Ki ± 0%   268.6Ki ± 0%   -0.12% (p=0.000 n=10)
Planner/large_cases.json-gen4left2right-20     174.7Ki ± 0%   174.4Ki ± 0%   -0.21% (p=0.000 n=10)
Planner/aggr_cases.json-gen4-20                7.183Mi ± 0%   7.215Mi ± 0%   +0.44% (p=0.000 n=10)
Planner/aggr_cases.json-gen4left2right-20      7.001Mi ± 0%   7.033Mi ± 0%   +0.45% (p=0.000 n=10)
Planner/select_cases.json-gen4-20              6.063Mi ± 0%   6.114Mi ± 0%   +0.84% (p=0.000 n=10)
Planner/select_cases.json-gen4left2right-20    5.958Mi ± 0%   6.010Mi ± 0%   +0.86% (p=0.000 n=10)
Planner/union_cases.json-gen4-20               2.215Mi ± 0%   2.230Mi ± 0%   +0.68% (p=0.000 n=10)
Planner/union_cases.json-gen4left2right-20     2.211Mi ± 0%   2.225Mi ± 0%   +0.66% (p=0.000 n=10)
SelectVsDML/DML_(random_sample,_N=32)-20       672.3Ki ± 0%   653.8Ki ± 0%   -2.75% (p=0.000 n=10)
SelectVsDML/Select_(random_sample,_N=32)-20   1155.5Ki ± 0%   973.9Ki ± 0%  -15.72% (p=0.000 n=10)
geomean                                        1.844Mi        1.804Mi        -2.18%

                                            │  ../before  │                after                │
                                            │  allocs/op  │  allocs/op   vs base                │
OLTP/Gen4-20                                  3.612k ± 0%   3.712k ± 0%   +2.77% (p=0.000 n=10)
OLTP/Gen4Greedy-20                            3.612k ± 0%   3.712k ± 0%   +2.77% (p=0.000 n=10)
OLTP/Gen4Left2Right-20                        3.612k ± 0%   3.712k ± 0%   +2.77% (p=0.000 n=10)
TPCC/Gen4-20                                  25.70k ± 0%   26.56k ± 0%   +3.36% (p=0.000 n=10)
TPCC/Gen4Greedy-20                            25.70k ± 0%   26.56k ± 0%   +3.36% (p=0.000 n=10)
TPCC/Gen4Left2Right-20                        25.45k ± 0%   26.31k ± 0%   +3.39% (p=0.000 n=10)
TPCH/Gen4-20                                  134.0k ± 0%   134.7k ± 0%   +0.55% (p=0.000 n=10)
TPCH/Gen4Greedy-20                            134.0k ± 0%   134.7k ± 0%   +0.55% (p=0.000 n=10)
TPCH/Gen4Left2Right-20                        124.1k ± 0%   124.7k ± 0%   +0.52% (p=0.000 n=10)
Planner/from_cases.json-gen4-20               106.7k ± 0%   107.5k ± 0%   +0.71% (p=0.000 n=10)
Planner/from_cases.json-gen4left2right-20     101.5k ± 0%   102.2k ± 0%   +0.75% (p=0.000 n=10)
Planner/filter_cases.json-gen4-20             562.7k ± 0%   476.1k ± 0%  -15.40% (p=0.000 n=10)
Planner/filter_cases.json-gen4left2right-20   560.7k ± 0%   474.1k ± 0%  -15.45% (p=0.000 n=10)
Planner/large_cases.json-gen4-20              10.52k ± 0%   10.54k ± 0%   +0.17% (p=0.000 n=10)
Planner/large_cases.json-gen4left2right-20    6.206k ± 0%   6.224k ± 0%   +0.29% (p=0.000 n=10)
Planner/aggr_cases.json-gen4-20               164.8k ± 0%   165.6k ± 0%   +0.52% (p=0.000 n=10)
Planner/aggr_cases.json-gen4left2right-20     160.2k ± 0%   161.1k ± 0%   +0.54% (p=0.000 n=10)
Planner/select_cases.json-gen4-20             140.4k ± 0%   142.1k ± 0%   +1.20% (p=0.000 n=10)
Planner/select_cases.json-gen4left2right-20   137.4k ± 0%   139.1k ± 0%   +1.22% (p=0.000 n=10)
Planner/union_cases.json-gen4-20              52.06k ± 0%   52.50k ± 0%   +0.85% (p=0.000 n=10)
Planner/union_cases.json-gen4left2right-20    51.89k ± 0%   52.33k ± 0%   +0.85% (p=0.000 n=10)
SelectVsDML/DML_(random_sample,_N=32)-20      12.55k ± 0%   13.97k ± 0%  +11.34% (p=0.000 n=10)
SelectVsDML/Select_(random_sample,_N=32)-20   26.53k ± 0%   21.93k ± 0%  -17.33% (p=0.000 n=10)
geomean                                       46.06k        45.76k        -0.65%
```

## Related Issue(s)

<!-- List related issues and pull requests. If this PR fixes an issue, please add it using Fixes #????  -->

## Checklist

-   [ ] "Backport to:" labels have been added if this change should be back-ported to release branches
-   [ ] If this change is to be back-ported to previous releases, a justification is included in the PR description
-   [ ] Tests were added or are not required
-   [ ] Did the new or modified tests pass consistently locally and on CI?
-   [ ] Documentation was added or is not required

## Deployment Notes

<!-- Notes regarding deployment of the contained body of work. These should note any db migrations, etc. -->