@frouioui 2024-08-29
> This Pull Request is now handled by arewefastyet. The current HEAD and future commits will be benchmarked.

I would not necessarily expect anything significant out of the benchmarks, query planning is a fraction of the overall performance.